Disallow conflicting sources
Different sources can assert facts about the same entity. This is a good thing.
However, if those sources assert different values for the same attribute, we
don't know which source takes presedence.

Prior to this commit, which value you saw was non-deterministic.
After this commit, an exception is thrown instead.